modularize app
