+++ src/proxy/mod.rs
@@ -1,0 +1,6 @@
+pub mod error;
+pub mod log;
+mod service;
+pub mod slurp;
+
+pub use service::Service;
